Isolate SMT lowerer
