+++ xdsl_smt/semantics/comb_semantics.py
@@ -13,7 +13,7 @@ from xdsl_smt.dialects import hw_dialect as hw
 from xdsl_smt.dialects import smt_bitvector_dialect as smt_bv
 from xdsl_smt.dialects import smt_dialect as smt
 from xdsl_smt.dialects import smt_utils_dialect as smt_utils
-from xdsl_smt.passes.lower_to_smt import SMTLowerer
+from xdsl_smt.passes.lower_to_smt.smt_lowerer import SMTLowerer
 from xdsl_smt.semantics.builtin_semantics import IntegerAttrSemantics
 from xdsl_smt.semantics.semantics import OperationSemantics
 from xdsl_smt.semantics.arith_semantics import SimplePurePoisonSemantics
